+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 - Add [ImageURLFilter](https://github.com/huacnlee/html-pipeline/blob/master/image_url_filter.go) match `img` to replace with rules like ([imageproxy](https://github.com/willnorris/imageproxy), Ban URL, Thumb version ...).
+- Add [ExternalLinkFilter](https://github.com/huacnlee/html-pipeline/blob/master/external_link_filter.go) a filter to match external links to add `rel="nofollow"`, `target="_blank"`.
 - Deprecated **ImageProxyFilter**, please use ImageURLFilter.
 
 ## 0.5.0